+++ tests/Feature/Expectations/HeaderNotExistsExpectationTest.php
@@ -46,4 +46,52 @@ class HeaderNotExistsExpectationTest extends TestCase
             $this->assertSame(404, $response->getStatusCode());
         }
     }
+
+    /**
+     * @test
+     */
+    public function itMatchesWhenHeaderIsMissing()
+    {
+        $builder = new MockBuilder();
+        $builder
+            ->when()
+                ->headerNotExist('Content-Type')
+            ->then()
+                ->body('bar');
+
+        $mock = new HttpMock($builder);
+
+        $client = new GuzzleClient();
+        $response = $client
+            ->withHandler($mock)
+            ->call('POST', 'https://example.com/v2/token');
+
+        $this->assertSame('bar', $response->getBody());
+    }
+
+    /**
+     * @test
+     */
+    public function itDoesNotMatchWhenHeaderIsPresent()
+    {
+        $builder = new MockBuilder();
+        $builder
+            ->when()
+                ->headerNotExist('Content-Type')
+            ->then()
+                ->body('bar');
+
+        $mock = new HttpMock($builder);
+
+        $client = new GuzzleClient();
+        $client
+            ->withHandler($mock)
+            ->prepareRequest('POST', 'https://example.com/v2/token')
+            ->getRequest()
+            ->setHeader('Content-Type', 'application/json');
+        $response = $client->execute();
+
+        $this->assertSame(404, $response->getStatusCode());
+        $this->assertSame('header \'Content-Type\' is present', $response->getBody());
+    }
 }